first-commit

+++ frontend/src/components/WorkoutDetails.jsx
@@ -1,16 +1,21 @@
 import { useWorkoutsContext } from "../hooks/useWorkoutContext";
-
+import { useAuthContext } from "../hooks/useAuthContext";
 // date function
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
 
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext();
+  const { user } = useAuthContext();
 
   const deleteHandler = async () => {
+    if (!user) {
+      return;
+    }
     const response = await fetch(
       "http://localhost:4000/api/workouts/" + workout._id,
       {
         method: "DELETE",
+        headers: { Authorization: `Bearer ${user.token}` },
       }
     );
     const data = await response.json();
